requirements.txt
requests
pillow
pytz
beautifulsoup4
